longest-dict-word-string-subsequence: Slight optimization
Sort the dictionary words in order of descending length and take
the first word that is a subsequence of string S.

*Measurements*
Original: 37 iterations
New: 23 iterations (excluding time to sort the array)

*From the website*
The running time is O(N*W) where W is the number of words in D
and N is the number of characters in S.

O(N*W) is asymptotically optimal if most dictionary words are
close to size N (because then the size of the input is O(N*W)).
However, it is far from optimal in a worst-case scenario where S
may be long and the dictionary strings quite short.

Define L to be the total number of characters in the dictionary
over all words. Then the best possible time complexity (assuming
words shorter than S) is O(N + L). However, the existing algorithm
is O(N * W), which might be O(N * L) in the case where dictionary
words are a small constant in size.